worker_us/requirements.txt

openbb>=4.5.0
openbb-yfinance>=1.3.0
openbb-cboe>=1.3.0
openbb-fred>=1.4.0
openbb-finviz>=1.3.0

pandas>=2.2.0

sqlalchemy>=2.0.35

python-dotenv>=1.0.0
tqdm>=4.66.0